feat(version): detect document version
Try to detect document version. Use (in the future) different trasformers, as v3 is a little bit
different. For now throw exception if version is not 2 and focus on the version 2 only

BREAKING CHANGE: It won't process document with newer syntax, it requires now to have the document
version specified